# Fix UAF issues with PCRE after request shutdown