--- a/ext/pcre/tests/gh15205_2.phpt
+++ b/ext/pcre/tests/gh15205_2.phpt
@@ -1,0 +1,30 @@
+--TEST--
+GH-15205: UAF when destroying stream after pcre request shutdown
+--CREDITS--
+nicolaslegland
+--FILE--
+<?php
+// Prime cache
+preg_replace('/pattern/', 'replace', 'subject');
+
+class wrapper
+{
+	public function stream_open($path, $mode, $options, &$opened_path)
+	{
+		return true;
+	}
+
+	public function stream_close()
+	{
+		echo "Close\n";
+		preg_replace('/pattern/', 'replace', 'subject');
+	}
+
+	public $context;
+}
+
+stream_wrapper_register('wrapper', 'wrapper');
+$handle = fopen('wrapper://', 'rb');
+?>
+--EXPECT--
+Close