@nielsdos 2024-08-03
I made a script that fills the cache with request-allocated strings:
```php
<?php
$f = fopen('in.php', 'w');
fwrite($f, '<?php'."\n");
for ($counter = 0; $counter < 10000; $counter++)
	fwrite($f, '\preg_match("/'.str_repeat("a", 100).$counter.'/", "subject");'."\n");
fclose($f);
```

Then I ran `out.php` with and without this patch (php_new is this PR):

```
Benchmark 1: ./sapi/cli/php in.php
  Time (mean ± σ):     124.7 ms ±   2.7 ms    [User: 115.5 ms, System: 8.6 ms]
  Range (min … max):   122.5 ms … 134.1 ms    24 runs
 
Benchmark 2: ./sapi/cli/php_new in.php
  Time (mean ± σ):     126.2 ms ±   2.1 ms    [User: 117.4 ms, System: 8.1 ms]
  Range (min … max):   124.7 ms … 134.5 ms    23 runs
 
Summary
  ./sapi/cli/php in.php ran
    1.01 ± 0.03 times faster than ./sapi/cli/php_new in.php
```

Callgrind numbers:
before this PR: 1,250,625,712
after this PR: 1,263,046,201
~0.98% difference