@iluuu1994 2024-08-03
@nielsdos The benchmark output says you ran in.php, just making sure that's not by accident? Also, wouldn't the biggest difference here come from using non-literal strings? I.e. strings constructed at runtime in the out.php script? Because they would have to be reallocated as persistent strings.