@nielsdos 2024-08-03
@iluuu1994 

> The benchmark output says you ran in.php, just making sure that's not by accident?

I meant to write `in.php`, not `out.php`, in the description, but the measurements are right. It's late :-)

> Also, wouldn't the biggest difference here come from using non-literal strings? I.e. strings constructed at runtime in the out.php script? Because they would have to be reallocated as persistent strings.

I'm running without opcache, so the strings are not persistent, and the check is against the `IS_STR_PERMANENT`. So we really are reallocating.

I previously tried to bench something like this:
```php
<?php
for ($counter = 0; $counter < 10000; $counter++)
	\preg_match("/" . str_repeat("a", 100) . "$counter/", "subject");
```
which will certainly not be persistent even with opcache, but then I realised that the string concat overhead and for loop overhead will be counted too. So I wanted to test a scenario where I had no such overhead so that the percentage difference is the most interpretable, which is why I ended up with the test script I posted in my previous comment.