@iluuu1994 2024-08-03
> I'm running without opcache, so the strings are not persistent, and the check is against the IS_STR_PERMANENT. So we really are reallocating.

Ah ok, thanks for the clarification. The 1% for this very synthetic benchmark shouldn't enough of a reason to bother with a different solution. Optimally, we could allocate interned strings as persistent in CLI while counting them towards the memory limit, which would avoid the realloc. But that's not something to do here.