@iluuu1994 2024-08-05
The uaf is not on the hash table itself, but on the `pcre_cache_entry`. But it might indeed be possible to use the existing refcount in `php_free_pcre_cache()` instead of always assuming the memory can be freed?