@nielsdos 2024-08-06
> The uaf is not on the hash table itself, but on the `pcre_cache_entry`. But it might indeed be possible to use the existing refcount in `php_free_pcre_cache()` instead of always assuming the memory can be freed?

Maybe, I investigated this a bit a month ago or so, so I don't remember the details but it got a bit messy to have both hash table destruction and refcount-based destruction because when the refcount reaches zero we shouldn't remove the cache entry otherwise the cache would almost always be empty. So it requires more changes than just that.