@nielsdos 2024-09-10
@iluuu1994 I fixed the remaining leak issue and updated the test to check for this. May you please have a final look?