@nielsdos 2024-09-11
Merged as ded8fb79bdce40bdb4c7ebd0a5dc91de0c9539a0. I'll let it sit in master and see to backport it (modulo struct changes) later when it has had time to sit and wait.